Add files via upload
The code estimates a constant from a series of nosy measurements of that constant using Least Squares Estimation. For example, we might have a temperature sensor attached to an engine and we would want to estimate the temperature of the engine. Each temperature measurement is very noisy since we have only a cheap sensor. So we will take multiple measurements with the sensor to try to get a better estimate of the true temperature.